Trekk ut noen av de større metodene i VedtakBegrunnelseProdusent.kt til egne filer
